Done scripts various usage

+++ scripts.py
@@ -32,37 +32,76 @@ TOP_COMMENDATIONS = [
 ]
 
 
-def fix_marks(schoolkid):
+def get_schoolkid(schoolkid_name):
+    from datacenter.models import Schoolkid
+    from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+    try:
+        return Schoolkid.objects.get(full_name__contains=schoolkid_name)
+    except MultipleObjectsReturned:
+        print(f"Слишком много совпадений по имени: {schoolkid_name}, уточните запрос.")
+        return
+    except ObjectDoesNotExist:
+        print(f"Пользователь с именем: {schoolkid_name} не найден.")
+        return
+
+
+def get_subject(subject_name, schoolkid):
+    from datacenter.models import Subject
+    from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+    from random import choice
+    if subject_name:
+        try:
+            return Subject.objects.get(title=subject_name, year_of_study=schoolkid.year_of_study)
+        except MultipleObjectsReturned:
+            print(f"Слишком много совпадений с предметом: {subject_name}, уточните запрос.")
+            return
+        except ObjectDoesNotExist:
+            print(f"Предмет с названием: {subject_name} не найден.")
+            return
+    else:
+        return choice(Subject.objects.filter(year_of_study=schoolkid.year_of_study))
+
+
+def fix_marks(schoolkid_name="Фролов Иван"):
     from datacenter.models import Mark
+    schoolkid = get_schoolkid(schoolkid_name)
+    if not schoolkid:
+        return
     frolov_bad_marks = Mark.objects.filter(schoolkid=schoolkid, points__lt=4)
     for mark in frolov_bad_marks:
         mark.points = 5
         mark.save()
 
 
-def remove_chastisements(schoolkid):
+def remove_chastisements(schoolkid_name="Фролов Иван"):
     from datacenter.models import Chastisement
+    schoolkid = get_schoolkid(schoolkid_name)
+    if not schoolkid:
+        return
     Chastisement.objects.filter(schoolkid=schoolkid).delete()
 
 
 def create_commendation(schoolkid_name="Фролов Иван", subject_name=None):
-    from datacenter.models import Commendation, Lesson, Subject, Schoolkid
-    from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+    from datacenter.models import Commendation, Lesson
     from random import choice
-    try:
-        schoolkid = Schoolkid.objects.get(full_name__contains=schoolkid_name)
-    except MultipleObjectsReturned:
-        print("Слишком много совпадений по введённому имени, уточните запрос.")
+    schoolkid = get_schoolkid(schoolkid_name)
+    if not schoolkid:
         return
-    except ObjectDoesNotExist:
-        print("По введённому имени пользователь не найден.")
+    subject = get_subject(subject_name, schoolkid)
+    if not subject:
         return
-    if subject_name:
-        subject = Subject.objects.get(title=subject_name, year_of_study=schoolkid.year_of_study)
-    else:
-        subject = choice(Subject.objects.filter(year_of_study=schoolkid.year_of_study))
-    schoolkid_lessons = Lesson.objects.filter(year_of_study=schoolkid.year_of_study, group_letter=schoolkid.group_letter, subject=subject)
+    schoolkid_lessons = Lesson.objects.filter(
+        year_of_study=schoolkid.year_of_study,
+        group_letter=schoolkid.group_letter,
+        subject=subject
+    )
     random_lesson = choice(schoolkid_lessons)
     commendation_text = choice(TOP_COMMENDATIONS)
-    Commendation.objects.create(text=commendation_text, created=random_lesson.date, schoolkid=schoolkid, subject=random_lesson.subject, teacher=random_lesson.teacher)
+    Commendation.objects.create(
+        text=commendation_text,
+        created=random_lesson.date,
+        schoolkid=schoolkid,
+        subject=random_lesson.subject,
+        teacher=random_lesson.teacher
+    )
